Move buildQuestion function to utility
