+++ src/main/java/com/google/sps/classes/Utility.java
@@ -17,9 +17,10 @@ package com.google.sps.classes;
 import com.google.appengine.api.users.UserService;
 import com.google.appengine.api.users.UserServiceFactory;
 import com.google.gson.Gson;
+import com.google.sps.classes.SqlConstants;
 import java.sql.*;
 import java.util.logging.Level;
-import java.util.logging.Logger;
+import java.util.logging.Logger; 
 
 // TODO(aabundis): Add JUnit tests for utility functions.
 
@@ -124,5 +125,29 @@ public final class Utility {
       Logger logger = Logger.getLogger(Utility.class.getName());
       logger.log(Level.SEVERE, exception.getMessage(), exception);
     }
+  }
+
+  /** 
+   * Create a question object using the results from a query.
+   */
+  private QuestionObject buildQuestion(ResultSet queryResult) {
+    QuestionObject question = new QuestionObject();
+    try {
+      question.setTitle(queryResult.getString(SqlConstants.QUESTION_FETCH_TITLE_COLUMN));
+      question.setBody(queryResult.getString(SqlConstants.QUESTION_FETCH_BODY_COLUMN));
+      question.setAskerId(queryResult.getInt(SqlConstants.QUESTION_FETCH_ASKERID_COLUMN));
+      question.setAskerName(queryResult.getString(SqlConstants.QUESTION_FETCH_AKSERNAME_COLUMN));
+      question.setDateTime(queryResult.getTimestamp(SqlConstants.QUESTION_FETCH_DATETIME_COLUMN));
+      question.setNumberOfFollowers(queryResult.getInt(
+          SqlConstants.QUESTION_FETCH_NUMBEROFFOLLOWERS_COLUMN));
+      question.setNumberOfAnswers(queryResult.getInt(
+          SqlConstants.QUESTION_FETCH_NUMBEROFANSWERS_COLUMN));
+    } catch (SQLException exception) {
+      // If the connection or the query don't go through, we get the log of what happened.
+      Logger logger = Logger.getLogger(FetchForumServlet.class.getName());
+      logger.log(Level.SEVERE, exception.getMessage(), exception);
+    }
+    
+    return question;
   }
 }
